+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v1.0.1-hotfix2
+- Made the "Use spawned cars in game" checkbox more visible and enabled by default
+- Added an Info tab in settings to further explain custom car/topper installation
+- Improved some tooltips in the UI
+
 ## v1.0.1-hotfix
 - Added a refresh button for the car dropdown in the RL Cars tab
     - Should fix the issue of no cars showing up (maybe)
@@ -9,4 +14,3 @@
 
 ## v1.0.0
 - First release 🥳
-